fix: reorg shopify code
base class got way too big and ugly.
