+++ shopify/backend/src/api/discount/discount-routes.ts
@@ -28,15 +28,13 @@ export async function create(request: IRequest, env: Env): Promise<Response> {
   Throw.ifNull(claims.dest);
 
   const shopify = new Shopify(claims.dest as string, env);
-  const rsp: DiscountRsp = {
-    id: await shopify.saveDiscount(body),
-  };
   const install = await shopify.getInstall();
-  await shopify.setDiscountActive(
-    install.data.currentAppInstallation.id,
-    rsp.id
-  );
-
+  const rsp: DiscountRsp = {
+    id: await shopify.createDiscount(
+      body,
+      install.data.currentAppInstallation.id
+    ),
+  };
   return json(rsp);
 }
