+++ shopify/backend/src/shopify/discount/shopify-discount-create.ts
@@ -3,10 +3,10 @@
  * MIT license. See LICENSE file in root directory.
  */
 
-import { ShopifyMetafield } from './shopify-metafield';
+import { ShopifyMetafield } from '../meta/shopify-metafield';
 import { ShopifyCombine } from './shopify-combine';
 
-export interface ShopifyDiscount {
+export interface ShopifyDiscountCreate {
   combinesWith: ShopifyCombine;
   endsAt?: Date;
   functionId: string;
