+++ shopify/backend/src/shopify/webhook/shopify-webhook-req-webhook.ts
@@ -3,7 +3,7 @@
  * MIT license. See LICENSE file in root directory.
  */
 
-export default interface ShopifyWebhook {
+export default interface ShopifyWebhookReqWebhook {
   address: string;
   topic: string;
   format: 'json' | 'xml';
