+++ shopify/backend/src/shopify/webhook/shopify-webhook.ts
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+import { ShopifyAuth } from '../auth/shopify-auth';
+import { API } from '@mytiki/worker-utils-ts';
+import { ShopifyWebhookReq } from './shopify-webhook-req';
+
+export class ShopifyWebhook extends ShopifyAuth {
+  async registerWebhook(webhook: ShopifyWebhookReq): Promise<void> {
+    const accessToken = await this.getToken();
+    await fetch(`https://${this.shopDomain}/admin/api/2023-04/webhooks.json`, {
+      method: 'POST',
+      headers: new API.HeaderBuilder()
+        .accept(API.Consts.APPLICATION_JSON)
+        .content(API.Consts.APPLICATION_JSON)
+        .set(ShopifyAuth.tokenHeader, accessToken)
+        .build(),
+      body: JSON.stringify(webhook),
+    });
+  }
+
+  registerOrderPaidWebhook = (baseUrl: string): Promise<void> =>
+    this.registerWebhook({
+      webhook: {
+        address: `https://${baseUrl}/api/latest/order/paid`,
+        topic: 'orders/paid',
+        format: 'json',
+      },
+    });
+}
